requirements.txt
Django
PyJWT
asgiref
djangorestframework
djangorestframework-simplejwt
patterns
pip
psycopg2
pytz
setuptools
sqlparse